+++ WinDirStat/windirstat/mainframe.cpp
@@ -293,7 +293,7 @@ void WDSSplitterWnd::SetSplitterPos( _In_ const DOUBLE pos ) noexcept {
 		if ( m_pColInfo == nullptr ) {
 			return;
 			}
-		const auto cxLeft = static_cast< INT >( pos * ( static_cast<std::int64_t>(rcClient.right) - static_cast<std::int64_t>(rcClient.left) ) );
+		const auto cxLeft = static_cast< INT >( pos * static_cast<DOUBLE>( static_cast<std::int64_t>(rcClient.right) - static_cast<std::int64_t>(rcClient.left) ) );
 		if ( cxLeft >= 0 ) {
 			CSplitterWnd::SetColumnInfo( 0, cxLeft, 0 );
 			CSplitterWnd::RecalcLayout( );
@@ -305,7 +305,7 @@ void WDSSplitterWnd::SetSplitterPos( _In_ const DOUBLE pos ) noexcept {
 	if ( m_pRowInfo == nullptr ) {
 		return;
 		}
-	const auto cyUpper = static_cast< INT >( pos * (static_cast<std::int64_t>(rcClient.bottom) - static_cast<std::int64_t>(rcClient.top) ) );
+	const auto cyUpper = static_cast< INT >( pos * static_cast<DOUBLE>(static_cast<std::int64_t>(rcClient.bottom) - static_cast<std::int64_t>(rcClient.top) ) );
 	if ( cyUpper >= 0 ) {
 		CSplitterWnd::SetRowInfo( 0, cyUpper, 0 );
 		CSplitterWnd::RecalcLayout( );
